server/project/requirements.txt
django
djangorestframework
django-cors-headers
fuzzywuzzy
pandas
python-Levenshtein